Bug 787502 - post: Move src.main.java.org.mozilla.gecko.sync.net.Server11* to src.main.java.org.mozilla.gecko.sync.net.server11.
