+++ src/main/java/org/mozilla/gecko/sync/net/server11/SyncServer11RequestDelegate.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.gecko.sync.net;
+package org.mozilla.gecko.sync.net.server11;
 
 public interface SyncServer11RequestDelegate {
   String credentials();
