+++ src/test/java/org/mozilla/android/sync/net/test/TestMetaGlobal.java
@@ -26,7 +26,7 @@ import org.mozilla.gecko.sync.MetaGlobal;
 import org.mozilla.gecko.sync.NonObjectJSONException;
 import org.mozilla.gecko.sync.delegates.MetaGlobalDelegate;
 import org.mozilla.gecko.sync.net.BaseResource;
-import org.mozilla.gecko.sync.net.SyncServer11Response;
+import org.mozilla.gecko.sync.net.server11.SyncServer11Response;
 import org.simpleframework.http.Request;
 import org.simpleframework.http.Response;
