+++ src/test/java/org/mozilla/android/sync/test/helpers/MockSyncClientsEngineStage.java
@@ -7,7 +7,7 @@ import static org.junit.Assert.assertTrue;
 
 import org.mozilla.gecko.sync.GlobalSession;
 import org.mozilla.gecko.sync.net.BaseResource;
-import org.mozilla.gecko.sync.net.SyncServer11Response;
+import org.mozilla.gecko.sync.net.server11.SyncServer11Response;
 import org.mozilla.gecko.sync.stage.SyncClientsEngineStage;
 
 public class MockSyncClientsEngineStage extends SyncClientsEngineStage {
